update the snack bar in util to show above bottom sheet too, UI issue fixed according to different mobile resolution in Task management screen, Punch History Screen, change text string class, fix bug in punch controller regarding the incorrect snackbar error, add a network timeout in network client class, fix box constraint in login screen, change data type of total days in leave response and leave repository. create Withdraw and update leave API and update the UI , add splash screen logo to fix the bug in launcher background

+++ lib/features/data/models/leave/leave_response.dart
@@ -12,7 +12,7 @@ class LeaveResponse {
   final DateTime appliedDate;
   final String? managerComment;
   final DateTime? processedDate;
-  final int totalDays;
+  final double totalDays;
   final String? shortLeaveTime;
   final String? halfDayType;
   final bool isNewlyApplied;
@@ -49,7 +49,7 @@ class LeaveResponse {
     DateTime? appliedDate,
     String? managerComment,
     DateTime? processedDate,
-    int? totalDays,
+    double? totalDays,
     String? shortLeaveTime,
     String? halfDayType,
     bool? isNewlyApplied,
